All non-abstract classes from Kununu\Elasticsearch\Result namespace are now final
