+++ src/Query/Criteria/Search/MultiFieldTrait.php
@@ -10,7 +10,7 @@ trait MultiFieldTrait
         $prepared = [];
         foreach ($fields as $key => $value) {
             if (is_array($value)) {
-                $boost = isset($value['boost']) ? '^' . $value['boost'] : '';
+                $boost = isset($value['boost']) ? sprintf('^%s', $value['boost']) : '';
                 $prepared[] = $key . $boost;
             } else {
                 $prepared[] = $value;
