+++ src/Result/AggregationResultSet.php
@@ -3,43 +3,10 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Result;
 
-/** @phpstan-consistent-constructor */
-class AggregationResultSet implements AggregationResultSetInterface
+final class AggregationResultSet extends AbstractAggregationResultSet
 {
-    protected ?ResultIteratorInterface $documents = null;
-    protected array $aggregationResults = [];
-
-    public function __construct(array $rawResult = [])
+    public static function create(array $rawResult = []): self
     {
-        foreach ($rawResult as $aggregationName => $fields) {
-            $this->aggregationResults[$aggregationName] = AggregationResult::create($aggregationName, $fields);
-        }
-    }
-
-    public static function create(array $rawResult = []): AggregationResultSet
-    {
-        return new static($rawResult);
-    }
-
-    public function setDocuments(ResultIteratorInterface $resultIterator): AggregationResultSetInterface
-    {
-        $this->documents = $resultIterator;
-
-        return $this;
-    }
-
-    public function getDocuments(): ?ResultIteratorInterface
-    {
-        return $this->documents;
-    }
-
-    public function getResultByName(string $name): ?AggregationResultInterface
-    {
-        return $this->aggregationResults[$name] ?? null;
-    }
-
-    public function getResults(): array
-    {
-        return $this->aggregationResults;
+        return new self($rawResult);
     }
 }
